backend/requirements.txt
deepface==0.0.80
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.9.0
firebase_admin==6.2.0
openmeteo_requests==1.3.0
pandas==2.0.3
# pyrebase==3.0.27
# urllib3=2.5
Pyrebase4==4.8.0
python-dotenv==1.0.1
# pytz==2023.3.post1
# requests==2.11.1
requests_cache==1.2.1
retry_requests==2.0.0
